requirements.txt
Django==3.2.25
djangorestframework==3.15.1
django-filter==23.5
djangorestframework-simplejwt==5.3.0
djangorestframework-api-key==2.3.0
django-cors-headers==4.1.0
importlib-metadata==6.7.0
PyJWT==2.8.0
pytz==2024.1
pillow==9.5.0